refactor(test): move old test case into test_legacy
