+++ test_legacy/exe_testcases/test_basic_name_be__main__/src/main.py
