+++ test_legacy/exe_testcases/test_sys_exit/src/main.py
