+++ test_legacy/test_exe.py
@@ -17,9 +17,10 @@ from __future__ import unicode_literals
 
 import os
 import subprocess
-from test import base
-from test.utility.gen_code_tools import lib_gen_pye
-from test.utility.pyconcrete_builder import pyconcrete_in_test_builder
+
+from test_legacy import base
+from test_legacy.utility.gen_code_tools import lib_gen_pye
+from test_legacy.utility.pyconcrete_builder import pyconcrete_in_test_builder
 
 
 class TestPyconcreteExe(base.TestPyConcreteBase):
